rename: ios/gpui_app.rs → ios/app.rs
Match the android/app.rs naming convention.
