+++ crates/zedra/src/ios.rs
@@ -1,2 +1,2 @@
 pub(crate) mod bridge;
-pub mod gpui_app;
+pub mod app;
